tenpy_bench: merge algorithm scripts into their test files
Each script's run_one(chi, L) now lives directly in its test_<name>.py
file instead of a separate, non-test module imported via a
sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..")) hack.
test_dmrg.py (which combined the dense and U(1)-symmetric DMRG cases)
is replaced by test_dmrg_dense.py and test_dmrg_symmetric.py, one file
per algorithm. tdvp.py (which, despite its name, runs TEBD internally
for cross-library consistency, per its own module docstring) becomes
test_tdvp.py.

variational_manual_grad.py's initial MPS, built via
MPS.from_random_unitary_evolution, now seeds numpy's RNG
(np.random.seed(0)) immediately beforehand, making the reached energy
after a fixed number of gradient steps reproducible run to run; without
it, the previous single-fixed-point regression check needed a wide
rel=1e-2 tolerance to absorb run-to-run variation in the unseeded
starting point.

Every benchmark test now asserts the returned energy against a
REFERENCE_ENERGIES dict covering the full (chi, L) grid, replacing the
previous single fixed-point regression check plus an isfinite-only
full-grid sweep. The grid is parametrized with two stacked
@pytest.mark.parametrize("chi", ...) / @pytest.mark.parametrize("length", ...)
decorators instead of a single parametrize over a combined (chi, length)
list, and limit_memory is applied as a plain decorator on a dedicated,
parameter-free *_memory test pinned to the (chi=16, L=20) point instead
of being attached via pytest.param(..., marks=...).
